+++ NEWS.rst
@@ -37,6 +37,11 @@ Bug fixes
   command specified by "I8C_CC" would be used even if "I8C_AS" was
   set.
 
+* I8X now fails gracefully if referenced symbol names cannot be
+  determined.  This can happen if an Infinity function provided by
+  a shared library loads an address that does not match any symbol
+  in the symbol table.
+
 Testsuite changes
 ~~~~~~~~~~~~~~~~~